Refactor - User(Auth)
DTO - class와 record 였던 것들 record 로 통합
반복되는 메시지 처리 import static 으로 코드 간소화
인터페이스 -> 구현체 구조, 인터페이스 제거 통일
에러코드 - 주석 추가
