+++ src/main/java/com/melog/melog/domain/user/dto/request/UserInfoDto.java
@@ -7,17 +7,5 @@ import lombok.NoArgsConstructor;
 
 import java.awt.print.Pageable;
 
-@Data
-@NoArgsConstructor
-public class UserInfoDto {
-    private String name;
-    private String email;
-    private Integer age;
-
-    @Builder
-    public UserInfoDto(User user) {
-        this.name = user.getName();
-        this.email = user.getEmail();
-        this.age = user.getAge();
-    }
-}
+public record UserInfoDto(String name, String email, Integer age)
+{}
