+++ src/main/java/com/melog/melog/domain/user/entity/User.java
@@ -14,7 +14,8 @@ public class User extends BaseTimeEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    @Column(name = "user_id")
+    private Long userId;
 
     @Column(nullable = false, unique = true, length = 50)
     private String email; //아이디로 활용
